Improve fetching client hint cache for the given origin
Client hints (ContentSettingsType::CLIENT_HINTS) are stored as per origin in PrefProvider, which is one of WebsiteSettings. Currently, to
fetch client hints for the given origins, we first fetch all client
hints rules and then do the linear scan to match the client hint rule
for the target origin, this is inefficiency and introduced unnecessary
latency. We change to call GetWebsiteSetting directly to read the client
hints cache for the given origin.

Bug: 1406891
Change-Id: Ie17fda42bcd4c49c02589481d4b6e2cdea7931d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4192109
Commit-Queue: Victor Tan <victortan@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098037}