Make the renderable fourcc list a parameter to VideoDecoderPipeline
Now that linux HW decode should be supporting a format that cros does
not, we need to separate this list based on platform. Additionally,
linux supports different textures based on the gl implementation,
so we need to configure this in a place that has access to the gl
info structures, hence, GpuMojoMediaClient(Cros).

Bug: 1399897
Change-Id: If7b704fcde23862069a2deaf77edaa4b5a4a2272
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4093896
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098079}