[dPWA] Fix cursor and right click behavior on App Home UI.
This CL does 2 things:
1. Ensures that hovering over icons on the UI changes the cursor to
a pointer.
2. Ensures that right clicking on other apps causes the context menu
for other apps to open, and right clicking anywhere else on the page
opens the default context menu.

Doing 2 was significantly harder due to CrActionMenu being a modal
that covers the entire dom and does not listen to any events except for one. To achieve 2, the following had to be done:
1. Move the right click context menu logic to be per app-item.
2. Have a series of events that are triggered from the app-item during
a right click on an app item (or when a context menu closes).
3. The app-list triggers the context menu closing, and also listens to
right click events on the document to handle non app use cases.

This CL also adds tests to test both of the above behaviors.

Bug: b/265025392, crbug.com/1350406
Change-Id: I66134ccd28392854ee2cc3f61bc8394de87da118
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4194053
Reviewed-by: Camden King <camdenking@google.com>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Phillis Tang <phillis@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098084}