[Read Anything] Add UKM metrics for Read Anything feature.
This will enable us to track the URLs where users are making the most
use of the Read Anything feature to aid in improving the underlying
models. The UKM source ID of the main frame of the active web contents
is passed from ReadAnythingController to ReadAnythingAppController. It
is passed to the AXTreeDistiller whenever the
ReadAnythingAppController requests distillation, and the
AXTreeDistiller forwards it to the ScreenAIServiceImpl which records
it when distillation completes.

The data flow of the UKM source ID is:
|-----------------------------|
| Browser process             |
|   ReadAnythingController    |
|   ReadAnythingModel         |
|   ReadAnythingPageHandler   |
|---- read_anything.mojom ----|
| Renderer process            |
|   ReadAnythingAppController |
|   AXTreeDistiller           |
|-- screen_ai_service.mojom --|
| Service process             |
|   ScreenAIServiceImpl       |
|-----------------------------|

UKM review doc:
https://docs.google.com/document/d/1bu443hg5ecH8_FwWdRWDv3w27jnX7aDHaMHiSCWhsRI/edit?pli=1&resourcekey=0-suFhHSowozm-dgv8J2D3Pg#heading=h.bahq24lmowo2

Bug: 1266555
Change-Id: I12e94e9a7571ca2d709cf832368e664881e6fb37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4152798
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Yue Ru Sun <yrsun@chromium.org>
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Roger McFarlane <rogerm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098102}