[CrOS Cellular] Fix sending empty UserAPNList to Shill
Fix error occurring when sending an empty Cellular.UserAPNList to Shill.
This was happening because Chrome was sending Shill an empty String List
rather than an empty Dictionary List. This is fixed by updating
shill_client_helper to explicitly check if the property being updated is
Cellular.UserAPNList, and if so, sending the correct list type.

Bug: b:162365553
Test: DUT
Change-Id: If631be3b4036889df9e3ba0322d3b80bb4b5abd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4199581
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Gordon Seto <gordonseto@google.com>
Cr-Commit-Position: refs/heads/main@{#1098111}