[Android] Fix a crash at premature fullscreen exit
Fixes a regression in fullscreen when it is stopped at pending
state. A remedy for a similar problem is present when an exit request
is received _before_ browser controls are hidden. This CL handles
the case where a request is receive _after_ the controls are hidden.
Makes sure the controls are shown again at exit, and fixes an NPE.

Bug: 1408997
Change-Id: Icbd1ae5a9b01ee72b25602643b9afc7d96ecad36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4199789
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098157}