Reland: PiP 2.0: Assume the feature is enabled on the browser side
This is a reland of crrev.com/c/4198715. The CL had a merge issue where
another test was added in a separate CL that landed simulataneously and
caused a compilation issue. The old description is copied below:

Due to the way origin trials work, on the browser side we need to just
assume that the origin trial is enabled on the renderer that requests a
new document pip window.

Bug: 1410552
Change-Id: Icbc27fb1c591fa97275d2a6ff9f71f3acbae422a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4200080
Auto-Submit: Tommy Steimel <steimel@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Commit-Queue: Mike Wasserman <msw@chromium.org>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098185}