[NTP][Panorama] Show which category is currently selected
Adds customize-chrome-check-mark-wrapper around each category to add a border and check mark to the category/type of the currently selected theme.

There is no selection state for the chrome web store tile so that is the only tile on the page that was left out of this change.

Also, added FocusOutlineManager so that the focus outline only shows with keyboard navigation.

screenshots:
large icon: screenshot/C3hsFmnSWZusWj3
small icon: screenshot/4qfishgTQpLm6TB

Bug: 1403334, 1402972
Change-Id: I928709d7d7c1de039264db6bc8081a5d9613886c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4196336
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Paul Adedeji <pauladedeji@google.com>
Commit-Queue: Riley Tatum <rtatum@google.com>
Cr-Commit-Position: refs/heads/main@{#1098216}