background_fetch: Migrate DOMException to `RejectWith-`
When throwing an exception `e`, it's important to make a stack trace
available as `e.stack` as long as possible.  However, DOMException's
constructor doesn't create this `stack` property, so the direct use of
DOMException's constructor is highly discouraged. So this patch migrates
`DOMException` constructor to `ScriptPromiseResolver::RejectWith-`.

Bug: 991544
Change-Id: I7619b0d839698d4ba7a13ff935ec940c5f5e5e06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4119817
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Jinho Bang <zino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098355}