Check Drive Quota and local volume space before copy or move in
When we call statvfs() on drivefs, it always returns to us that it has 1
exobyte of space available, so we currently always allow a copy or move
operation into Drive. This can be a problem when either we don't have
enough local disk, or we are out of Drive quota.

We now check both.

Bug: b/266163411
Change-Id: I542e636308d6e02f7910ac6054981762dd1d4bb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4191625
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098433}