DLP: Disable blocked share sheet apps
The CL enables graying out of blocked apps in the Share sheet, needed
for DLP files restrictions.

To do this, the file manager passes the source urls, kept in the file's
metadata, to AppService which uses it to check the DLP status of each
task as it's listing them. Information about blocked tasks is then
passed to the SharesheetService which can disable the button/icon for
such tasks (apps).

http://shortn/_h2ct4Y48YM

Fixed: b/265798535
Change-Id: I2e39d306233d8faa4498bb9f3439a9b70723c959
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4173661
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Commit-Queue: Aida Zolic <aidazolic@chromium.org>
Reviewed-by: Wenbo Jie <wenbojie@chromium.org>
Reviewed-by: Melissa Zhang <melzhang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098509}