[GPU] Eliminate external access to Texture::ImageState
In practice, clients simply need to know whether there is a level image
that needs binding present. This CL replaces direct client access to
Texture::ImageState with the ability to make that query, adapts
clients, and makes |Texture::LevelInfo::image_state| private to Texture.

This change also enforces the fact that there are no flows that query
the ImageState when the image is null (verifying that the value of the
ImageState doesn't actually matter in that case).

Bug: 1323341
Change-Id: Ic8ed195e123a634d95ea4e984c373f59471cbe1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4194847
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098516}