Fix a crash when using OptionalBooleanToBool on undefined value
if we type in a field which has not been autofilled, |was_autofilled| is
undefined, so |filled_value_was_modified| is also undefined. For
|filled_value_was_modified| we should set it in both filling and
typing cases, we want to set it to be true when filling happens first
and the user types into the field, if autofill happens without a user
editing to it, it should be false.

Bug: 1410898
Change-Id: I6c37b82c02fb85d7f647ff88ca4d795b3b7b986d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4201202
Commit-Queue: Dominic Battré <battre@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098530}