[SafetyCheck] Add undo for 'Allow again' action for unused sites module
When clicking 'Allow again' for an origin's revoked permissions in the unused sites module the permission(s) is(are) granted again, the item is removed from the list of revoked permissions, and the origin is added to an ignore list for the future. This CL adds the undo functionality on the CPP side (TS side already landed).

Bug: 1345920, 1409416
Change-Id: If88b1a7d5a988dabfa0260c1725b4d1998523fed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4176863
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Illia Klimov <elklm@chromium.org>
Commit-Queue: Eduard Hez <ehez@chromium.org>
Reviewed-by: Rainhard Findling <rainhard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098533}