DLP: Add integration test for blocked Arc(VM)
Adds an integration test that checks Play files are disabled when added
as the ARC container or as the ARCVM (initially placeholder than real).

Fixes disabling of Play files when they're added as a placeholder,
before the VM is actually mounted, and of the real volume that's added
when the VM is mounted (though this shouldn't happen).

Bug: b/263072872
Change-Id: I6a02e71f9a2ff19ac32d2fa649c86c4706727306
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4161537
Reviewed-by: Wenbo Jie <wenbojie@chromium.org>
Commit-Queue: Aida Zolic <aidazolic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098547}