[IOS][Omnibox] Add an a11y identifier for images.
Add an a11y identifier to Omnibox favicon/suggestion
type image view.

There is no user-visible change.

This can be used in EG Tests to verify that the correct
icon is displayed.

The identifier is different depending on the state:

For default search engine icon, the identifier is
"kOmniboxLeadingImageDefaultAccessibilityIdentifier"

For empty omnibx: the identifier is
"kOmniboxLeadingImageEmptyTextAccessibilityIdentifier"

For favicon : the identifier is the URL of the web page for
which the favicon is shown.

For suggestion type icons :
the identifier is the string representing the match type.

Bug: 1405433
Change-Id: Ie75931616b173227706393134d5fe30b762adcd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4180846
Commit-Queue: Ameur Hosni <ameurhosni@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098550}