HighlightInheritance: fix perf regression in UA ::target-text rule
CL:3546467 adds a rule for a:-webkit-any-link::target-text to the UA
style sheet for HTML (html.css).

Under highlight inheritance (experimental as of CL:4001367), when we
recalc styles for any link, the rule forces us to also recalc styles
for the link’s ::target-text, causing large regressions in CSS perf
tests containing links.

Since the underline declaration was only needed as a workaround for
pre-overlay highlight painting (stable in M105), and the background
declaration was in turn needed as a workaround for pre-inheritance
highlight styling, this patch removes the rule entirely.

We also remove the tests that were added in CL:3546467, because they
assert the old non-conformant behaviour.

Results for perf_tests/css/FocusUpdate.html:

• 24920 26088 runs/s, before, HighlightInheritance off
• 18744 19924 runs/s, before, HighlightInheritance on
• 32652 29506 runs/s, after, HighlightInheritance off
• 32155 28578 runs/s, after, HighlightInheritance on

Results for perf_tests/css/LoadBootstrapBlog.html:

• 6.11 6.01 6.08 ms, before, HighlightInheritance off
• 7.20 6.86 7.00 ms, before, HighlightInheritance on
• 6.34 6.40 5.98 ms, after, HighlightInheritance off
• 6.26 6.45 6.36 ms, after, HighlightInheritance on

Fixed: 1408712, 1408764
Change-Id: I32c41e05866b7912801557975a829dc42db7cf3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4199622
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Delan Azabani <dazabani@igalia.com>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098555}