[GPU] Make ImageState::NOIMAGE for when there is no image in Texture
Currently Texture::LevelInfo::image_state is UNBOUND when there is no
GLImage in the Texture (i.e., Texture::LevelInfo::image is null). This
CL adds a new NOIMAGE state for that case. There is no client-visible
impact as no clients see the state directly; rather, this just adds
clarity to the code and additionally adds the invariant that the
image state can be set to UNBOUND only by SetUnboundLevelImage(), which
is useful.

Bug: 1323341
Change-Id: Icd9cecb3cc311c67141df04f7fb336fcf6bd8d82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4197085
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098575}