[Autofill] Add UMA to track usage of "Manage addresses..." suggestion
This CL adds a new enumeration histogram,
Autofill.SuggestionsListManageClicked, that is logged when the
"Manage..." entry is selected by a user on autofill pop-up
menu. The histogram is only recorded for Payment Method,
Address, and Personal Info dialogs - the Passwords dialog takes
a different code path and is not in scope for what
we are trying to measure here. (Clicking on
"Manage Passwords..." is measured via
PasswordManager.PasswordDropdownItemSelected histogram)

The purpose of this histogram is to get a rough estimate of the
frequency that users use the "Manage..." links for each supported
data type (Passwords excluded), primarily as a gut-check for
whether or not these links are used.

Note that the new histogram is not normalized for the popularity
of different types of Autofill, and as such care should be taken
when interpreting it. (For example, by manually normalizing the
results against
`Autofill.FormEvents.[Address|CreditCard].Suggestions`).



BUG=1406348

Change-Id: I4a0e3380be00ca900b6bcdf5747668dc4146193c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4150206
Reviewed-by: Matthias Körber <koerber@google.com>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Vidhan Jain <vidhanj@google.com>
Commit-Queue: Peter Kotwicz <pkotwicz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098623}