wifi: Update default creation timestamp returned
This change updates the default timestamp returned when the creation
timestamp for a WiFi network is requested from the NetworkMetadataStore.
Previously, the current timestamp was returned in the case when a
network had not existed long enough to be considered for migration.
However, this will only be accurate the first time it is returned.

Bug: b/236148589
Test: Unit tests.
Change-Id: Id5bd4ea982aad4c15c2cc5e01397967f2ad61b51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4191001
Reviewed-by: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Chad Duffin <chadduffin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098643}