Handle attribution reporting response header in fenced frame beacons
https://github.com/WICG/turtledove/blob/main/Fenced_Frames_Ads_Reporting.md#support-for-attribution-reporting

Fenced frame beacons are sent in the browser process, therefore are
not observed by the resource observer in blink.

Instead of adding a observer in the browser process, fenced frame
will notify ARA when a beacon has been sent and when a response
data has been received as it has access to all the data.

Sources registered in the navigation beacon are blocked until the
navigation completed successfully and will be dropped if the
navigation failed.

Fenced frame reporting is expected to integrate with AttributionHost
as follows:

1. AttributionHost::NotifyFencedFrameReportingBeaconSent is called
when the attribution eligible beacon is sent. For navigation beacon
it should be called before the navigation commit in order to access
the initiator frame.

2. AttributionHost::NotifyFencedFrameReportingBeaconData is called
when the redirect response or the final response is received for the
beacon, within the callbacks of SimpleURLLoader::SetRedirectCallback
and SimpleURLLoader::DownloadHeadersOnly.

We expect that a unique identifier is sent for each beacon, for
navigation beacon the identifier should be the navigation id in order
to associate with the navigation.

Bug: 1408902
Change-Id: Id00d4287a2b6a7cb2d7459776904e6f79756684b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4182443
Commit-Queue: Nan Lin <linnan@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098679}