[StorageKey] Add *_if_third_party_enabled to mojom
This CL expands the StorageKey mojom definition to include the
*_if_third_party_enabled fields alongside the "normal" fields.
It also checks the incoming data to see if it would make a well formed
StorageKey.

Previously when serializing a StorageKey we would use the values of
these *_if_third_party_enabled fields and then infer on the receiving
side what the "normal" fields should be.

Doing this however makes it more difficult to correctly determine a
StorageKey's fields once it depends on RuntimeFeatureState's results.

By making these fields explicit it's easier to check whether the key is
well formed and makes it simpler to check correctness afterward.

Bug: 1199077, 1377000
Change-Id: If7f91c5ff118d7c35e887749b6410a28569ebf96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4192118
Reviewed-by: Ari Chivukula <arichiv@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098683}