[TTFCC] Start the bottom sheet in full height for accessibility
When TalkBack is enabled, the bottom sheet will default to an expanded
state to make all actions accessible to the screen reader users.
Going back to the half state on swiping down is disabled if a screen
reader is in use.

This change applies to the bottom sheets for passwords and payments.

This change is required in the green lines for Touch to Fill for
Payments.

The reason for the change is that a blind user has no benefit from the
half state since there is nothing that the bottom sheet is covering for
them. It's also allowing them to navigate the whole sheet content.

Bug: 1247698
Change-Id: If8336afa72446cac210a6338c862b6a15568b1ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4188378
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#1098703}