[CBCM] Allows ProxyPolicyProvider to own the delegate provider
This CL fixes an inverted pointer ownership dependency in CBCM code:
- ChromeBrowserPolicyConnector owns a pointer to
  MachineLevelUserCloudPolicyManager
- ChromeBrowserPolicyConnector's base class owns a ProxyPolicyProvider
  whose delegate is the MachineLevelUserCloudPolicyManager object owned
  by its subclass

The consequence of that inverted dependency is that the base class'
destructor may try to reference a dangling pointer deleted by the
derived class' destructor. That might happen in some special early
shutdown situations when proper pointer cleanup is not performed.

The issue is resolved by allowing the ProxyPolicyProvider object to
optionally own the delegate object. We explored moving ownership for
all use cases, but some of them are very trick - e.g. [1] in which not
only can the delegate object be multiple different objects, but also
they are exposed via API and used in many places that might outlive
the proxy object.

[1] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/policy/profile_policy_connector.cc;l=265;drc=71630a0e336a703e21de9ebeb98a5abf84e8c96c

Fixed: 1398421
Bug: 1398421
Change-Id: Ia676d84e0ad05f6db77b536be4e031ee34cacbb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4171927
Commit-Queue: Fabio Tirelo <ftirelo@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098746}