Refactor PaintOp skip
This is mainly to resolve my own confusion about 'skip` :)

We used 'skip' to indicate the size of memory space for a PaintOp. It
had two meanings:
1. The aligned size of a PaintOp in PaintOpBuffer;
2. The serialized size of a PaintOp in serialization buffer.
A PaintOp may have different byte format in PaintOpBuffer and in
serialized buffer, and the values of the two skips may be very
different.

Now rename #1 to aligned_size, and #2 to serialized_size.

Bug: 1385848
Change-Id: Iad245481da23b38ff9263676ce1be9188480f48a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4193404
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098748}