media/gpu/v4l2/test: Remove unique_ptr from H264DPB.
The H264 Decoder had previously setup the DPB as a vector of unique
pointers. While this works, it is unnecessary as the H264SliceMetaData
elements in the DPB could just be used as raw data.

This update modifies the H264DPB class to work as a map of
H264SliceMetaData objects instead of unique pointers to the slice data.
This update modifies the corresponding decoder functions to no longer
use unique_ptr for the slice data.

BUG=b:234752983
TEST=v4l2_stateless_decoder --video=BA1_Sony_D.h264 --md5=checksums.log
TEST=v4l2_stateless_decoder --video=BAMQ1_JVC_C.h264 --md5=checksums.log
TEST=v4l2_stateless_decoder --video=CABA1_SVA_B.h264 --md5=checksums.log

Change-Id: I46677d09f86715eb91d2c65255cda1584ee137df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4195676
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Steve Cho <stevecho@chromium.org>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1098764}